feat: enhance backend with robust operations, duplicates resolution, plugins, and error handling
- **Undo/Redo Stack**: Refactored `OperationHistory` in `file_operations.py` to use JSON serialization instead of `pickle` for secure persistence to `~/.tfm/history.json`. Added `to_dict` and `from_dict` methods to `FileOperation`.
- **Duplicate Resolution**: Enhanced `cli.py` to support `largest` and `smallest` strategies, mapping to `ConflictResolutionStrategy`. Included live progress bar updates for duplicate resolutions.
- **Error Handling & Logging**: Replaced generic `except:` and `except Exception:` blocks with specific exception catching (`OSError`, `json.JSONDecodeError`) across `ai_mode.py`, `ai_integration.py`, and `file_operations.py`. Used structured logging instead of `pass` or `print`.
- **Plugin System**: Verified the `PluginRegistry` successfully loads from `~/.tfm/plugins/` and hooks into operations like `on_file_added` and `on_organize`.
- **Testing Expansion**: Added and verified comprehensive tests using `tmp_path` fixtures for `OperationHistory`, `PluginRegistry`, `FileOrganizer`, `FileSearcher`, and `FileOperations`. All 93 tests pass successfully.
- **Documentation**: Updated `USER_MANUAL.md` reflecting changes to JSON storage, new duplicate resolution CLI options, and plugin examples.

Linting and type checking (`ruff`, `mypy`) pass completely.

+++ src/file_manager/context.py
@@ -5,7 +5,7 @@ Directory Context Builder for AI Prompts.
 import os
 import time
 from pathlib import Path
-from typing import Dict, Any, List, Optional
+from typing import Dict, Any
 from dataclasses import dataclass, asdict
 
 @dataclass
@@ -90,7 +90,7 @@ class DirectoryContextBuilder:
             newest_file=newest_name
         )
 
-    def _human_size(self, size: int) -> str:
+    def _human_size(self, size: float) -> str:
         for unit in ['B', 'KB', 'MB', 'GB', 'TB']:
             if size < 1024.0:
                 return f"{size:.1f} {unit}"
